fix(picker): telescope dependency check
Was testing out the picker functionality and found it odd that I was
being told I didn't have telescope installed. Sure enough, I found the
check in the code was requiring an extension.

Because I could not find the extension mentioned anywhere in the readme,
I decided to remove it and just run the check against telescope on it's
own instead. This worked just fine :)